backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib[bcrypt]
bcrypt==3.2.2
pydantic
python-jose[cryptography]
python-multipart
